Changed names of folders to correspond to what the folders do, not the author
