+++ .travis.yml
@@ -2,6 +2,6 @@ language: python
 install:
   - pip install coveralls
 script:
-  nosetests --with-coverage AlexDanoff Azeem Joon
+  nosetests --with-coverage tex2Wiki DLMF_preprocessing Joon
 after_success:
   coveralls
