+++ DLMF_preprocessing/src/remove_excess.py
