+++ Pipeline/src/pipeline.py
@@ -1,14 +1,14 @@
 import glob
 import sys
 
-sys.path.append('../../AlexDanoff/src')
+sys.path.append('../../DLMF_preprocessing/src')
 from utilities import readin as readin
 from utilities import writeout
 from remove_excess import remove_excess as step1
 from replace_special import remove_special as step2
 from prepare_annotations import prepare_annotations as step3
 
-sys.path.append('../../Azeem/src')
+sys.path.append('../../tex2Wiki/src')
 from tex2Wiki import readin as step4
 from tex2Wiki import writeout as step5
