+++ macro_replacement/cosSubstitution.py
