+++ macro_replacement/function_regex
