+++ tex2Wiki/test/test_isnumber.py
